MT#61408 Don't fail if there is no revisions to delete
This is rollback case if during the upgrade there is no
not_replicated sql scripts.
In 1st run of 'ngcp-update-db-schema --downgrade-release' deletes all
the records will be deleted on sp1. So the next run on sp2 finds no
records to delete and it's ok.
So don't fail on this case.

Change-Id: Ifcdd6cb369eaa0da0167c50023e5502e06aa3a8a