MT#61373 Update some columns for reproducible builds
Fix of db-schema.git:29ce6639d6a5101bf008274b678f022a449ca8e0 commit.

Instead of using string '1970-01-01 00:00:00' use
FROM_UNIXTIME(1).
Using the string causes the date '0000-00-00 00:00:00' and the warning:
+---------+------+-----------------------------------------------------+
| Level   | Code | Message                                             |
+---------+------+-----------------------------------------------------+
| Warning | 1264 | Out of range value for column 'applied_at' at row 1 |
+---------+------+-----------------------------------------------------+

It's because the time zones differences so '1970-01-01 00:00:00'
translates into earlier date which is out of range for timestamp
datatype.

Using the FROM_UNIXTIME() gives us a reliable result which fits the
datatype and is not zero date.

Change-Id: Ie123363a74e81c37c8ca5458b6aafd53fd3b74ca