MT#63431 Create new upscript to cope with changes done to 15418.up
The upscript 15418.up was created in 2018 and recently it was adapted
for debian/trixie. In turn, this will cause warnings in ngcp-status as
ngcp-mysql-compare-dbs script shows inconsistencies between the json
file and the db itself.

Original change:
719fd2750bf6d225c561e9011a07c329f7accd4f

Adaptations:
57b924cce02568381850b00f19f64d05fddb5e40
39e6d65add451999ad6317ecda6bb9be8e474e38
374f456ffb97a47378c2a33673fc0135efa4f951

To fix these inconsistencies, add a new upscript with the exact same
details as in 15418.up including the latest adaptations and a 'DROP
PROCEDURE IF EXISTS.

(cherry picked from commit 41769aa2a054ff523769abc92512465b5b95692e)

Change-Id: I9a3d2373954326b2f6786dc8c5c27fc58cb73c44